Merge branch 'cherry-pick-0b533d69' into 'CPG136_Release'

See merge request IvyProjects/IvyCpg_Product_MobileApp_Android!988

(cherry picked from commit fc2a9395a0f206564f9c84a9f4c4bbba60c03c99)

7f87f36c Merge branch 'cherry-pick-0b533d69' into 'CPG135_Release'

+++ ivycpg/src/main/java/com/ivy/cpg/view/denomination/DenominationFragment.java
@@ -251,7 +251,7 @@ public class DenominationFragment extends IvyBaseFragment {
                             textViewHashMap.get(key).setText("0");
                     }
 
-                    if (totalValues <= Integer.valueOf(initialTotalAmount))
+                    if (totalValues <= Double.valueOf(initialTotalAmount))
                         mTotalCollectionTextview.setText(getActivity().getResources().getString(R.string.total) + ":" + String.valueOf(totalValues));
                 }
             }
@@ -284,7 +284,7 @@ public class DenominationFragment extends IvyBaseFragment {
                         textViewHashMap.get(key).setText("0");
                 }
 
-                if (totalValues <= Integer.valueOf(initialTotalAmount))
+                if (totalValues <= Double.valueOf(initialTotalAmount))
                     mTotalCollectionTextview.setText(getActivity().getResources().getString(R.string.total) + ":" + String.valueOf(totalValues));
                 else
                     Toast.makeText(getActivity(), "" + getActivity().getResources().getString(R.string.denomination_error), Toast.LENGTH_SHORT).show();
